+++ packages/backend/src/schema.rs
@@ -9,7 +9,7 @@ diesel::table! {
         location_country -> Nullable<Text>,
         level -> Nullable<Text>,
         domains -> Nullable<Text>,
-        date_posted_timestamp_ms -> Nullable<Integer>,
-        date_updated_timestamp_ms -> Nullable<Integer>,
+        date_posted_timestamp_ms -> Nullable<BigInt>,
+        date_updated_timestamp_ms -> Nullable<BigInt>,
     }
 }
